--- Arriba/Arriba.Web/jsx/SearchHeader.scss
@@ -1,147 +1,0 @@
-.input { // Corrdinate styles between input and rail.
-    font-size: 16pt;
-    padding: 8px;
-}
-
-.header {
-    height: 64px;
-    display: flex;
-    flex-flow: row;
-
-    /* Top Left Title - white, middle, bigger */
-    .title {
-        cursor: default;
-        padding: 16px 8px;
-        vertical-align: middle;
-        font-size: 18pt;
-    }
-
-    /* The title is a link; fix it to look like plain text */
-    .title a {
-        color: white;
-        font-weight: normal;
-    }
-
-    .searchBarAndButtons {
-        flex: 1;
-        display: flex;
-        align-items: center;
-        padding: 8px;
-        padding-right: 10px;
-
-        & > a { // <a> assumed to be buttons.
-            padding: 8px;
-            font-size: 0; // Discard white space.
-            &:hover { opacity: 0.8; }
-        }
-    }
-
-    .searchBar {
-        background: hsl(0, 0, 97%);
-        flex: 1;
-        display: flex;
-        align-items: center;
-        position: relative;
-        padding-right: 4px;
-        margin-right: 8px;
-    }
-
-    /* Make the search box large, fill the space, well padded, slightly grey */
-    .searchBar input { @extend .input;
-        box-sizing: border-box;
-        border: none;
-        flex: 1;
-        color: #666666;
-        background: transparent; // Override user agent.
-    }
-
-    /* Make the search icon (magnifying glass) the same as the search box and right next to it */
-    .searchBar .searchIcon {
-        font-size: 16pt;
-        color: #666666;
-        padding: 4px;
-    }
-
-    /* The indefinite progress bar to when waiting for the server */
-    .searchBar .loading {
-        display: none;
-        position: absolute;
-        bottom: 0;
-        height: 3px;
-        width: 100%;
-
-        &.loading-active {
-            display: block;
-            background: linear-gradient(90deg, transparent 70%, hsl(136, 100%, 23%) 85%, transparent);
-            background-size: 200%;
-            animation: wave 1.5s linear 500ms infinite;
-
-            @keyframes wave {
-                0%    { background-position: 200% 50% }
-                100%  { background-position: 0%   50% }
-            }
-        }
-    }
-}
-
-// Contains invisible mirrored text to align the completion list.
-.rail {
-    display: inline-flex;
-    position: absolute;
-    left: 0; bottom: 0; height: 0;
-    white-space: pre;
-    visibility: hidden;
-    & > * { visibility: visible; }
-
-    @extend .input;
-    // Mimic global input styles manually (unable to programmatically reference).
-    font-family: "Segoe UI", "Segoe WP", Tahoma, Arial, sans-serif;
-    font-weight: 600;
-    padding-top: 0; padding-bottom: 0; // Override any possible padding values;
-}
-
-.suggestions {
-    position: absolute;
-    top: 2px; left: -8px; // Manually match input padding;
-    background-color: white;
-    box-shadow: 0 3px 8px 0 rgba(0,0,0,0.2), 0 0 0 1px rgba(0,0,0,0.08);
-    z-index: 1;
-    max-height: 70vh;
-    min-width: 150px;
-    overflow-y: auto;
-
-    svg {
-        position: absolute;
-        right: 0;
-        width: 80px;
-        height: 100%;
-        z-index: -1;
-        fill: hsla(130, 100%, 30%, 0.15);
-    }
-}
-
-.suggestions .suggestion {
-    font-size: 12pt;
-    font-weight: 600;
-    padding: 8px;
-    cursor: pointer;
-    box-sizing: border-box;
-    white-space: nowrap;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.suggestions .suggestion:hover {
-     background-color: rgba(0, 0, 0, 0.02); 
-}
-
-.suggestions .suggestion-sel {
-    background-color: hsla(216, 80%, 85%, 0.25)
-}
-
-.suggestions .suggestion-hint {
-    color: hsl(220, 40%, 50%);
-    margin-left: 30px;
-    font-size: 9pt;
-}
