+++ Arriba/Arriba.Web/parts.optional/WorkItemDetails/WorkItemHistory.scss
