+++ Arriba/Arriba.Web/parts/DefaultConfiguration.jsx
