+++ Arriba/Arriba.Web/parts/DetailsTable.jsx
