+++ Arriba/Arriba.Web/parts/ResultDetails.jsx
