+++ Arriba/Arriba.Web/parts/SearchHeader.jsx
