+++ Arriba/Arriba.Web/parts/SplitPane.jsx
